+++ inst/shiny/v1.4/immune_repertoire/server.R
@@ -527,6 +527,13 @@ source(
 source(
   paste0(
     Cerebro.options[["cerebro_root"]],
+    "/shiny/v1.4/immune_repertoire/precomputed.R"
+  ),
+  local = TRUE
+)
+source(
+  paste0(
+    Cerebro.options[["cerebro_root"]],
     "/shiny/v1.4/immune_repertoire/visualizations.R"
   ),
   local = TRUE